requirements.txt
asgiref==3.7.2
astroid==3.0.0
black==23.9.1
click==8.1.7
colorama==0.4.6
dill==0.3.7
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
isort==5.12.0
mccabe==0.7.0
mypy==1.6.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pylint==3.0.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
